refactor : 형식 인증 함수들 utils폴더에 따로 빼도록 구조 수정 / 더미 데이터 constants 폴더에 따로 관리하도록 수정
