+++ src/constants/dummyData.ts
@@ -1,0 +1,139 @@
+// 더미 데이터 따로 저장
+
+// 로그인 더미 데이터
+const dummyLoginInfo = [
+  {
+    userId: "kurum",
+    userPw: "1234",
+  },
+];
+
+// 인증코드 더미 데이터
+const dummyCode = "123456";
+// 유저들이 사용하고 있는 닉네임
+const dummyNicknames = ["asdfqwer", "kuroom", "asdf1234"];
+
+// 단과대 선택
+const colleges = [
+  "문과대학",
+  "이과대학",
+  "건축대학",
+  "공과대학",
+  "사회과학대학",
+  "경영대학",
+  "부동산과학원",
+  "KU융합과학기술원",
+  "상허생명과학대학",
+  "수의과대학",
+  "예술디자인대학",
+  "사범대학",
+  "상허교양대학",
+  "국제대학",
+  "융합과학기술원",
+  "생명과학대학",
+];
+
+// 학과 선택
+const departments = {
+  문과대학: [
+    "국어국문학과",
+    "영어영문학과",
+    "중어중문학과",
+    "철학과",
+    "사학과",
+    "지리학과",
+    "미디어커뮤니케이션학과",
+    "문화콘텐츠학과",
+    "문과대학자유전공학부",
+  ],
+  이과대학: ["수학과", "화학과", "물리학과", "이과대학자유전공학부"],
+  건축대학: ["건축학부"],
+  공과대학: [
+    "사회환경공학부",
+    "기계항공공학부",
+    "전기전자공학부",
+    "화학공학부",
+    "컴퓨터공학부",
+    "산업경영공학부 산업공학과",
+    "산업경영공학부 신산업융합학과",
+    "생물공학과",
+    "K뷰티산업융합학과",
+    "기계・로봇・자동차공학부",
+    "화공학부",
+    "첨단융합공학부 재료공학과",
+    "첨단융합공학부 항공우주・모빌리티공학과",
+    "첨단융합공학부 생물공학과",
+    "첨단융합공학부 산업공학과",
+    "공과대학자유전공학부",
+  ],
+  사회과학대학: [
+    "정치외교학과",
+    "경제학과",
+    "행정학과",
+    "국제무역학과",
+    "응용통계학과",
+    "융합인재학과",
+    "글로벌비즈니스학과",
+    "사회과학대학융합전공학부",
+  ],
+  경영대학: ["경영학과", "기술경영학과"],
+  부동산과학원: ["부동산학과"],
+  KU융합과학기술원: [
+    "미래에너지공학과",
+    "스마트운행체공학과",
+    "스마트ICT융합공학과",
+    "화장품공학과",
+    "줄기세포재생공학과",
+    "의생명공학과",
+    "시스템생명공학과",
+    "융합생명공학과",
+  ],
+  상허생명과학대학: [
+    "생명과학특성학과",
+    "동물자원과학과",
+    "식량자원과학과",
+    "축산식품생명공학과",
+    "식품유통공학과",
+    "환경보건과학과",
+    "산림조경학과",
+  ],
+  수의과대학: ["수의예과", "수의학과"],
+  예술디자인대학: [
+    "커뮤니케이션디자인학과",
+    "산업디자인학과",
+    "의상디자인학과",
+    "리빙디자인학과",
+    "현대미술학과",
+    "영상학과",
+    "매체연기학과",
+  ],
+  사범대학: [
+    "일어교육과",
+    "수학교육과",
+    "체육교육과",
+    "음악교육과",
+    "교육공학과",
+    "영어교육과",
+    "교직과",
+  ],
+  상허교양대학: ["KU자유전공학부", "교양교육센터", "국제학부"],
+  국제대학: ["국제통상학과", "문화미디어학과", "국제교양교육센터"],
+  융합과학기술원: [
+    "첨단바이오공학부",
+    "생명공학부 시스템생명공학과",
+    "생명공학부 융합생명공학과",
+    "융합과학기술원자유전공학부",
+  ],
+  생명과학대학: [
+    "동물자원・식품과학・유통학부 동물자원전공",
+    "동물자원・식품과학・유통학부 식품과학전공",
+    "동물자원・식품과학・유통학부 식품유통전공",
+    "환경보건・산림조경학부 환경보건전공",
+    "환경보건・산림조경학부 산림조경전공",
+    "생명과학특성학과",
+    "식량자원과학과",
+    "생명과학대학자유전공학부",
+  ],
+};
+
+export { dummyLoginInfo, dummyCode, dummyNicknames, colleges, departments };
